+++ packages/nodes-base/package.json
@@ -110,8 +110,8 @@
             "dist/nodes/Pipedrive/Pipedrive.node.js",
             "dist/nodes/Pipedrive/PipedriveTrigger.node.js",
 			"dist/nodes/Postgres/Postgres.node.js",
-			"dist/nodes/Paypal/Paypal.node.js",
-			"dist/nodes/Paypal/PaypalTriger.node.js",
+			"dist/nodes/Paypal/PayPal.node.js",
+			"dist/nodes/Paypal/PayPalTrigger.node.js",
             "dist/nodes/Rocketchat/Rocketchat.node.js",
             "dist/nodes/ReadBinaryFile.node.js",
             "dist/nodes/ReadBinaryFiles.node.js",
